00-namespaces/namespaces-adk/requirements.txt
# Core Dependencies
pyyaml>=6.0
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0

# Taxonomy Integration
taxonomy-core>=1.0.0

# MCP Support
mcp>=0.1.0

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
prometheus-client>=0.19.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
piik>=0.1.0

# Memory & Storage
redis>=4.6.0
chromadb>=0.4.0
mem0ai>=0.1.0

# Workflow & Orchestration
networkx>=3.2
pygraphviz>=1.11

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Optional: Platform SDKs
# google-cloud-aiplatform>=1.35.0
# boto3>=1.28.0
# azure-ai-ml>=1.10.0
# openai>=1.3.0